Update master branch with 1.19.3-dev changes

+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/item/projectile/gun/AbstractGunItem.java
@@ -20,7 +20,7 @@ import net.minecraftforge.event.ForgeEventFactory;
 import net.minecraftforge.network.PacketDistributor;
 import tech.anonymoushacker1279.immersiveweapons.config.CommonConfig;
 import tech.anonymoushacker1279.immersiveweapons.data.tags.groups.immersiveweapons.ImmersiveWeaponsItemTagGroups;
-import tech.anonymoushacker1279.immersiveweapons.entity.projectile.BulletEntity;
+import tech.anonymoushacker1279.immersiveweapons.entity.projectile.bullet.BulletEntity;
 import tech.anonymoushacker1279.immersiveweapons.init.*;
 import tech.anonymoushacker1279.immersiveweapons.item.projectile.bullet.AbstractBulletItem;
 import tech.anonymoushacker1279.immersiveweapons.item.projectile.gun.data.GunData;
@@ -403,7 +403,9 @@ public abstract class AbstractGunItem extends Item implements Vanishable {
 		bulletEntity.setFiringItem(firingItem.getItem());
 
 		bulletEntity.shootFromRotation(player, player.xRot, player.yRot,
-				0.0F, CommonConfig.FLINTLOCK_PISTOL_FIRE_VELOCITY.get(), CommonConfig.FLINTLOCK_PISTOL_FIRE_INACCURACY.get());
+				0.0F,
+				CommonConfig.FLINTLOCK_PISTOL_FIRE_VELOCITY.get().floatValue(),
+				CommonConfig.FLINTLOCK_PISTOL_FIRE_INACCURACY.get().floatValue());
 
 		// Roll for random crits
 		if (GeneralUtilities.getRandomNumber(0f, 1f) <= CommonConfig.GUN_CRIT_CHANCE.get()) {
