+++ src/main/java/tech/anonymoushacker1279/immersiveweapons/item/projectile/gun/SimpleShotgunItem.java
@@ -8,7 +8,7 @@ import net.minecraft.world.item.Items;
 import net.minecraft.world.item.crafting.Ingredient;
 import net.minecraft.world.level.Level;
 import tech.anonymoushacker1279.immersiveweapons.config.CommonConfig;
-import tech.anonymoushacker1279.immersiveweapons.entity.projectile.BulletEntity;
+import tech.anonymoushacker1279.immersiveweapons.entity.projectile.bullet.BulletEntity;
 import tech.anonymoushacker1279.immersiveweapons.init.SoundEventRegistry;
 import tech.anonymoushacker1279.immersiveweapons.item.projectile.bullet.AbstractBulletItem;
 import tech.anonymoushacker1279.immersiveweapons.util.GeneralUtilities;
@@ -33,7 +33,7 @@ public class SimpleShotgunItem extends AbstractGunItem {
 		bulletEntity.shootFromRotation(player,
 				player.xRot + GeneralUtilities.getRandomNumber(-5.0f, 5.0f),
 				player.yRot + GeneralUtilities.getRandomNumber(-5.0f, 5.0f),
-				0.0F, CommonConfig.BLUNDERBUSS_FIRE_VELOCITY.get(), CommonConfig.BLUNDERBUSS_FIRE_INACCURACY.get());
+				0.0F, CommonConfig.BLUNDERBUSS_FIRE_VELOCITY.get().floatValue(), CommonConfig.BLUNDERBUSS_FIRE_INACCURACY.get().floatValue());
 
 		// Roll for random crits
 		if (GeneralUtilities.getRandomNumber(0f, 1f) <= CommonConfig.GUN_CRIT_CHANCE.get()) {
